pyproject.toml
[project]
name = "fithub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.7",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "django-filter>=24.2",
    "django-cors-headers>=4.3.1",
    "drf-spectacular>=0.27.0",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-django>=4.8.0",
    "pytest-cov>=4.0.0",
    "factory-boy>=3.3.0",
    "faker>=22.0.0",
    "testcontainers>=3.7.0",
    "docker>=6.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "bandit>=1.7.0",
    "safety>=2.3.0",
]

[tool.black]
line-length = 127
target-version = ['py313']
exclude = '''
/(
    \.venv
    | venv
    | __pycache__
    | \.git
)/
'''

[tool.isort]
profile = "black"
line_length = 127
skip_glob = [".venv/*", "venv/*", "__pycache__/*", ".git/*"]
